+++ internal/feature/watchlist/domain/entity/user_symbol.go
@@ -6,8 +6,8 @@ import "time"
 // UserSymbol はユーザーのウォッチリストエントリを表します。
 // watchlists テーブルにマップされ、users.id と symbols.code に FK 制約を持ちます。
 type UserSymbol struct {
-	ID         uint
-	UserID     uint
+	ID         int64
+	UserID     int64
 	SymbolCode string
 	SortKey    int
 	CreatedAt  time.Time
