+++ internal/feature/watchlist/transport/handler/watchlist_handler.go
@@ -17,10 +17,10 @@ import (
 
 // WatchlistUsecase はウォッチリスト操作のユースケースインターフェースを定義します。
 type WatchlistUsecase interface {
-	ListUserSymbols(ctx context.Context, userID uint) ([]entity.UserSymbol, error)
-	AddSymbol(ctx context.Context, userID uint, symbolCode string) error
-	RemoveSymbol(ctx context.Context, userID uint, symbolCode string) error
-	ReorderSymbols(ctx context.Context, userID uint, orderedCodes []string) error
+	ListUserSymbols(ctx context.Context, userID int64) ([]entity.UserSymbol, error)
+	AddSymbol(ctx context.Context, userID int64, symbolCode string) error
+	RemoveSymbol(ctx context.Context, userID int64, symbolCode string) error
+	ReorderSymbols(ctx context.Context, userID int64, orderedCodes []string) error
 }
 
 // WatchlistHandler はウォッチリストに関連するHTTPリクエストを処理します。
@@ -35,7 +35,7 @@ func NewWatchlistHandler(uc WatchlistUsecase) *WatchlistHandler {
 
 // List はユーザーのウォッチリスト一覧を取得します。
 func (h *WatchlistHandler) List(c *gin.Context) {
-	userID := c.MustGet(jwtmw.ContextUserID).(uint)
+	userID := c.MustGet(jwtmw.ContextUserID).(int64)
 
 	entries, err := h.uc.ListUserSymbols(c.Request.Context(), userID)
 	if err != nil {
@@ -47,7 +47,7 @@ func (h *WatchlistHandler) List(c *gin.Context) {
 	out := make([]api.WatchlistItem, 0, len(entries))
 	for _, e := range entries {
 		out = append(out, api.WatchlistItem{
-			Id:         int(e.ID),
+			Id:         e.ID,
 			SymbolCode: e.SymbolCode,
 			SortKey:    e.SortKey,
 		})
@@ -57,7 +57,7 @@ func (h *WatchlistHandler) List(c *gin.Context) {
 
 // Add はウォッチリストに銘柄を追加します。
 func (h *WatchlistHandler) Add(c *gin.Context) {
-	userID := c.MustGet(jwtmw.ContextUserID).(uint)
+	userID := c.MustGet(jwtmw.ContextUserID).(int64)
 
 	var req api.AddWatchlistRequest
 	if err := c.ShouldBindJSON(&req); err != nil {
@@ -83,7 +83,7 @@ func (h *WatchlistHandler) Add(c *gin.Context) {
 
 // Remove はウォッチリストから銘柄を削除します。
 func (h *WatchlistHandler) Remove(c *gin.Context) {
-	userID := c.MustGet(jwtmw.ContextUserID).(uint)
+	userID := c.MustGet(jwtmw.ContextUserID).(int64)
 	code := c.Param("code")
 
 	if err := h.uc.RemoveSymbol(c.Request.Context(), userID, code); err != nil {
@@ -102,7 +102,7 @@ func (h *WatchlistHandler) Remove(c *gin.Context) {
 
 // Reorder はウォッチリストの並び順を更新します。
 func (h *WatchlistHandler) Reorder(c *gin.Context) {
-	userID := c.MustGet(jwtmw.ContextUserID).(uint)
+	userID := c.MustGet(jwtmw.ContextUserID).(int64)
 
 	var req api.ReorderWatchlistRequest
 	if err := c.ShouldBindJSON(&req); err != nil {
