+++ internal/feature/watchlist/usecase/watchlist_usecase.go
@@ -11,14 +11,14 @@ import (
 
 // WatchlistRepository はウォッチリスト操作の永続化層を抽象化します。
 type WatchlistRepository interface {
-	ListByUser(ctx context.Context, userID uint) ([]entity.UserSymbol, error)
+	ListByUser(ctx context.Context, userID int64) ([]entity.UserSymbol, error)
 	// Add はsort_keyを指定してウォッチリストに銘柄を追加します。
 	Add(ctx context.Context, entry entity.UserSymbol) error
 	// AddWithNextSortKey はsort_keyをトランザクション内でMAX+1採番して銘柄を追加します。
 	// MaxSortKey取得とInsertをアトミックに実行するため、並行追加時の重複順位を防ぎます。
-	AddWithNextSortKey(ctx context.Context, userID uint, symbolCode string) error
-	Remove(ctx context.Context, userID uint, symbolCode string) error
-	UpdateSortKeys(ctx context.Context, userID uint, entries []entity.UserSymbol) error
+	AddWithNextSortKey(ctx context.Context, userID int64, symbolCode string) error
+	Remove(ctx context.Context, userID int64, symbolCode string) error
+	UpdateSortKeys(ctx context.Context, userID int64, entries []entity.UserSymbol) error
 }
 
 // SymbolExistsChecker は銘柄の存在確認を行うインターフェースです。
@@ -40,14 +40,14 @@ func NewWatchlistUsecase(repo WatchlistRepository, symbolChecker SymbolExistsChe
 }
 
 // ListUserSymbols はユーザーのウォッチリストをソート順で返します。
-func (u *WatchlistUsecase) ListUserSymbols(ctx context.Context, userID uint) ([]entity.UserSymbol, error) {
+func (u *WatchlistUsecase) ListUserSymbols(ctx context.Context, userID int64) ([]entity.UserSymbol, error) {
 	return u.repo.ListByUser(ctx, userID)
 }
 
 // AddSymbol はウォッチリストに銘柄を追加します。
 // symbols テーブルに存在しない銘柄コードの場合は ErrSymbolNotFound を返します。
 // 既にウォッチリストに存在する場合は ErrAlreadyInWatchlist を返します。
-func (u *WatchlistUsecase) AddSymbol(ctx context.Context, userID uint, symbolCode string) error {
+func (u *WatchlistUsecase) AddSymbol(ctx context.Context, userID int64, symbolCode string) error {
 	exists, err := u.symbolChecker.Exists(ctx, symbolCode)
 	if err != nil {
 		return fmt.Errorf("checking symbol existence: %w", err)
@@ -60,12 +60,12 @@ func (u *WatchlistUsecase) AddSymbol(ctx context.Context, userID uint, symbolCod
 }
 
 // RemoveSymbol はウォッチリストから銘柄を削除します。
-func (u *WatchlistUsecase) RemoveSymbol(ctx context.Context, userID uint, symbolCode string) error {
+func (u *WatchlistUsecase) RemoveSymbol(ctx context.Context, userID int64, symbolCode string) error {
 	return u.repo.Remove(ctx, userID, symbolCode)
 }
 
 // ReorderSymbols はウォッチリストの並び順を更新します。
-func (u *WatchlistUsecase) ReorderSymbols(ctx context.Context, userID uint, orderedCodes []string) error {
+func (u *WatchlistUsecase) ReorderSymbols(ctx context.Context, userID int64, orderedCodes []string) error {
 	entries := make([]entity.UserSymbol, 0, len(orderedCodes))
 	for i, code := range orderedCodes {
 		entries = append(entries, entity.UserSymbol{
@@ -79,13 +79,13 @@ func (u *WatchlistUsecase) ReorderSymbols(ctx context.Context, userID uint, orde
 
 // OnUserCreated は PostSignupHook インターフェースを実装します。
 // サインアップ直後に呼ばれ、デフォルト銘柄を追加します。
-func (u *WatchlistUsecase) OnUserCreated(ctx context.Context, userID uint) error {
+func (u *WatchlistUsecase) OnUserCreated(ctx context.Context, userID int64) error {
 	return u.InitializeDefaults(ctx, userID)
 }
 
 // InitializeDefaults は新規ユーザー向けにデフォルト銘柄（AAPL/MSFT/GOOGL）を追加します。
 // symbols テーブルに存在する銘柄のみ追加します（存在しない場合はスキップ）。
-func (u *WatchlistUsecase) InitializeDefaults(ctx context.Context, userID uint) error {
+func (u *WatchlistUsecase) InitializeDefaults(ctx context.Context, userID int64) error {
 	defaultSymbols := []string{"AAPL", "MSFT", "GOOGL"}
 
 	for i, code := range defaultSymbols {
